+++ packages/react-form-composer/src/index.js
@@ -14,6 +14,7 @@ export {default as setStateValueByPath} from './state-utils/set-field';
 export {default as Scope, useScope } from './scope';
 export {default as FormStateContext} from './form-state-context';
 export {default as FormStateProvider} from './form-state-provider';
+export {default as reduxFormStateProvider} from './redux-form-state-provider';
 export {default as FormContextProvider} from './form-context-provider';
 export {default as formReducer} from './reducers';
 export {default as Field} from './field';
@@ -26,6 +27,14 @@ import {updateField} from './actions';
 
 export const updateFieldAction = (field, value) => (
   {field, ...updateField(value)}
+);
+
+export const updateFormFieldsAction = (form, payload) => (
+  {form, ...updateFields(payload)}
+);
+
+export const updateFormFieldAction = (form, field, value) => (
+  {form, field, ...updateField(value)}
 );
 
 export { Text } from './ui-components/text';