+++ packages/react-form-composer-redux-provider/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 No pending (unreleased changes) to mention
 
+## [2.0.1] - 2019-08-18
+### Fixed
+- Update README to explain that from version 2.6.0 of react-form-composer you do not need this package.
+
+## [2.0.1] - 2019-06-15
+### Fixed
+- Update react-form-composer peer dependency to v2.0.0.
+
+## [2.0.0] - 2019-06-15
+### Fixed
+- Update readme.
+
 ## [1.0.0] - 2019-06-13
 ### Added
 - Initial release. This is a rename/rebrand of [redux-formkit-redux-state-provider](https://www.npmjs.com/package/redux-formkit-redux-state-provider).